Revert League Detail Page to Previous State
